Back end database connect to user activity
Now when a user registers they have to enter shipping address etc. Linked to shipping_address table comfirmed. When user adds item to cart, the back-end functionality works so it shows up in the database the user has added an item and creates temporary order in the orders table which is 'pending', until checkout